Move ParsePciAddress function to hardware utils
