+++ pkg/util/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "hw_helper.go",
         "os_helper.go",
         "util.go",
     ],
