--- pkg/util/hw_helper.go
@@ -1,40 +1,0 @@
-/*
- * This file is part of the KubeVirt project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Copyright 2018 Red Hat, Inc.
- *
- */
-
-package util
-
-import (
-	"fmt"
-	"regexp"
-)
-
-const PCI_ADDRESS_PATTERN = `^([\da-fA-F]{4}):([\da-fA-F]{2}):([\da-fA-F]{2}).([0-7]{1})$`
-
-// ParsePciAddress returns an array of PCI DBSF fields (domain, bus, slot, function)
-func ParsePciAddress(pciAddress string) ([]string, error) {
-	pciAddrRegx, err := regexp.Compile(PCI_ADDRESS_PATTERN)
-	if err != nil {
-		return nil, fmt.Errorf("failed to compile pci address pattern, %v", err)
-	}
-	res := pciAddrRegx.FindStringSubmatch(pciAddress)
-	if len(res) == 0 {
-		return nil, fmt.Errorf("failed to parse pci address %s", pciAddress)
-	}
-	return res[1:], nil
-}
