+++ pkg/virt-api/webhooks/validating-webhook/admitters/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
     deps = [
         "//pkg/controller:go_default_library",
         "//pkg/hooks:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/util/webhooks:go_default_library",
         "//pkg/util/webhooks/validating-webhooks:go_default_library",
