+++ pkg/virt-api/webhooks/validating-webhook/admitters/vmi-create-admitter.go
@@ -35,8 +35,8 @@ import (
 
 	v1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/kubevirt/pkg/hooks"
-	"kubevirt.io/kubevirt/pkg/util"
 	"kubevirt.io/kubevirt/pkg/util/hardware"
+	hwutil "kubevirt.io/kubevirt/pkg/util/hardware"
 	webhookutils "kubevirt.io/kubevirt/pkg/util/webhooks"
 	"kubevirt.io/kubevirt/pkg/virt-api/webhooks"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
@@ -754,7 +754,7 @@ func ValidateVirtualMachineInstanceSpec(field *k8sfield.Path, spec *v1.VirtualMa
 		}
 		// verify that the specified pci address is valid
 		if iface.PciAddress != "" {
-			_, err := util.ParsePciAddress(iface.PciAddress)
+			_, err := hwutil.ParsePciAddress(iface.PciAddress)
 			if err != nil {
 				causes = append(causes, metav1.StatusCause{
 					Type:    metav1.CauseTypeFieldValueInvalid,
@@ -1712,7 +1712,7 @@ func validateDisks(field *k8sfield.Path, disks []v1.Disk) []metav1.StatusCause {
 				})
 			}
 
-			_, err := util.ParsePciAddress(disk.Disk.PciAddress)
+			_, err := hwutil.ParsePciAddress(disk.Disk.PciAddress)
 			if err != nil {
 				causes = append(causes, metav1.StatusCause{
 					Type:    metav1.CauseTypeFieldValueInvalid,
