+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//pkg/host-disk:go_default_library",
         "//pkg/ignition:go_default_library",
         "//pkg/util:go_default_library",
+        "//pkg/util/hardware:go_default_library",
         "//pkg/util/net/dns:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
